+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Changed signature of Driver. Technically a breaking change, but unlikely to affect anyone.
 
+### Added
+
+- Added `DataTable.remove_row` method https://github.com/Textualize/textual/pull/2253
+- `Widget.scroll_to_center` now scrolls the widget to the center of the screen https://github.com/Textualize/textual/pull/2255
+- Added `TabActivated` message to `TabbedContent` https://github.com/Textualize/textual/pull/2260
+
+
 ## [0.19.1] - 2023-04-10
 
 ### Fixed